asyncRequestInputStream failure case test
